# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Services Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "List all services",
                "operationId": "4c175c5d6594b4a5112bbe15a46e2bb4",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current Page",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order: latest or oldest",
                        "schema": {
                            "default": "latest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "create a new service",
                "operationId": "b05b71c30267faa468172458580fb6fc",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "customer_id"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "customer id",
                                        "default": ""
                                    },
                                    "name": {
                                        "description": "service name",
                                        "default": ""
                                    },
                                    "description": {
                                        "description": "service description",
                                        "default": ""
                                    },
                                    "price": {
                                        "description": "service price",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/customers/{customer_id}/services": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all services of a single customer",
                "operationId": "c1206c6fec02513886df7126f7d5afb7",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "customer_id",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current Page",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order: latest or oldest",
                        "schema": {
                            "default": "latest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/services/{service_id}": {
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "update service",
                "operationId": "d7c6ff10d18577d94b77657038ecc8ab",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "service id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "customer_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "service name",
                                        "default": ""
                                    },
                                    "description": {
                                        "description": "service description",
                                        "default": ""
                                    },
                                    "price": {
                                        "description": "service price",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "delete service",
                "operationId": "83305dc9bf75c006eee0fbd03327325c",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "service id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "should send the Authorization header with \"Bearer <jwt_token>\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}